% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread.R
\name{bread.lavaan}
\alias{bread.lavaan}
\title{Bread Matrix of the Sandwich Matrix}
\usage{
bread.lavaan(x, ...)
}
\arguments{
\item{x}{a fitted model object.}
}
\value{
Returns a matrix.
}
\description{
These functions extract an estimator for the bread of sandwiches.
}
