% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bread.R
\name{bread}
\alias{bread}
\title{Bread Matrix of the Sandwich Matrix}
\usage{
bread(x, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{...}{arguments passed to methods.}
}
\value{
Returns a matrix.
}
\description{
These functions extract an estimator for the bread of sandwiches.
}
