% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ipcr_arguments.R
\name{check_ipcr_arguments}
\alias{check_ipcr_arguments}
\title{Check the specified arguments of ipcr}
\usage{
check_ipcr_arguments(fit, iterated, conv, max_it, linear, model_data)
}
\arguments{
\item{fit}{a fitted model.}

\item{iterated}{a logical value.}

\item{conv}{an integer used as stopping criterion for iterated IPC regression.}

\item{max_it}{the maximum number of iterations for iterated IPC regressions.}

\item{linear}{a logical value.}
}
\description{
Internal function that checks if the arguments specified in the ipcr function are valid.
}
