% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.ipcr.R
\name{coef.ipcr}
\alias{coef.ipcr}
\title{Extract Individual Parameter Regression Estimates}
\usage{
\method{coef}{ipcr}(object, parameter = NULL, method = "iterated", ...)
}
\arguments{
\item{object}{an ipcr object.}

\item{parameter}{a character vector with target model parameters that are to be inspected.}

\item{method}{a character string indicating whether to show results iterated IPC regression (iterated) or standard IPC regression (standard).}
}
\value{
Returns a list.
}
\description{
This function extracts the individual parameter regression estimates from an ipcr object.
}
