% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.ipcr.R
\name{effects.ipcr}
\alias{effects.ipcr}
\title{Effects from Fitted Model}
\usage{
\method{effects}{ipcr}(object, parameter = NULL, method = "iterated", ...)
}
\arguments{
\item{object}{an ipcr object.}

\item{parameter}{a character vector with target model parameters that are to be inspected.}

\item{method}{a character string indicating whether to show results iterated IPC regression (iterated) or standard IPC regression (standard).}
}
\value{
Returns a list.
}
\description{
Return (orthogonal) effects from a fitted IPC regression model.
}
