% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estfun.R
\name{estfun}
\alias{estfun}
\title{Extract Empirical Estimating Functions}
\usage{
estfun(x, ...)
}
\arguments{
\item{x}{a fitted MxRAMModel-type model.
... further arguments passed to or from other methods.}
}
\value{
a n x q matrix corresponding to n observations and q parameters.
}
\description{
A function for extracting the empirical estimating functions of a fitted MxRAMModel-type model. This is the derivative of the objective function with respect to the parameter vector, evaluated at the observed (case-wise) data. In other words, this function returns the case-wise scores, evaluated at the fitted model parameters.
}
