% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.ipcr.R
\name{fitted.ipcr}
\alias{fitted.ipcr}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{ipcr}(object, parameter = NULL, method = "iterated", ...)
}
\arguments{
\item{object}{an ipcr object.}

\item{parameter}{a character vector with target model parameters that are to be inspected.}

\item{method}{a character string indicating whether to show results iterated IPC regression (iterated) or standard IPC regression (standard).}
}
\value{
Returns a list.
}
\description{
This function extracts the fitted values from IPC regression models.
}
