% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytic_jac.R
\name{get_analytic_jac}
\alias{get_analytic_jac}
\title{Calculate the Analyitc Jacobian Matrix}
\usage{
get_analytic_jac(x)
}
\arguments{
\item{x}{list with information about the model.}

\item{deriv}{list with partial derivatives of the RAM matrices.}
}
\value{
Jacobian matrix.
}
\description{
Internal function that anayltically calculates the Jacobian matrix.
}
