% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytic_jac_input.R
\name{get_analytic_jac_input}
\alias{get_analytic_jac_input}
\title{Prepare Input for the Analytic Gradient}
\usage{
get_analytic_jac_input(x)
}
\arguments{
\item{x}{list with information about the model.}
}
\value{
List with partial derivatives.
}
\description{
Internal function that calculates the partial derivative of the RAM matrices with respect to the parameters.
}
