% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytic_jac_up.R
\name{get_analytic_jac_up}
\alias{get_analytic_jac_up}
\title{Calculate the Analyitc Jacobian Matrix}
\usage{
get_analytic_jac_up(x, A_up, S_up, M_up, F_up, B_up = B_up)
}
\arguments{
\item{x}{list with information about the model.}

\item{deriv}{list with partial derivatives of the RAM matrices.}
}
\value{
Jacobian matrix.
}
\description{
Internal function that anayltically calculates the Jacobian matrix.
}
