% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arguments_checked.R
\name{get_arguments_checked}
\alias{get_arguments_checked}
\title{Check the specified arguments of ipcr}
\usage{
get_arguments_checked(fit, covariates, iterated, conv, max_it)
}
\arguments{
\item{fit}{a structural equation model estimated with OpenMx.}

\item{covariates}{a data frame or matrix with covariates.}

\item{iterated}{a logical value.}

\item{conv}{an integer used as stopping criterion for iterated IPC regression.}

\item{max_it}{the maximum number of iterations for iterated IPC regressions.}
}
\description{
Internal function that checks if the arguments specified in the ipcr function are valid.
}
