% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centered_moment_deviations.R
\name{get_centered_moment_deviations}
\alias{get_centered_moment_deviations}
\title{Calculate Centered Individual Deviations}
\usage{
get_centered_moment_deviations(x)
}
\arguments{
\item{x}{list with information about the model.}
}
\value{
Matrix with centered individual devations from the sample moments.
}
\description{
Internal function that calculates the individual deviations from the predicted individual- or group-specific mean and covariance.
}
