% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get the data used to fit the model}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{a fitted model object}
}
\value{
Returns a data frame.
}
\description{
This function gets the data used to fit the model and returns it as data frame.
}
