% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Partition the Samples Into Similar Groups}
\usage{
get_groups(x)
}
\arguments{
\item{x}{list with information about the model.}

\item{md}{matrix with individual deviations from the sample moments.}
}
\value{
Matrix with groupings.
}
\description{
Internal function that partitions the samples in groups with the same values of covariates.
}
