% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initial_IPCs.R
\name{get_initial_IPCs}
\alias{get_initial_IPCs}
\title{Calculate Initial Individual Parameter Contributions}
\usage{
get_initial_IPCs(x)
}
\arguments{
\item{x}{list with information about the model.}
}
\value{
Matrix of initial individual parameter contributions.
}
\description{
Internal function that calculates the initial parameter contributions.
}
