% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_info.R
\name{get_model_info}
\alias{get_model_info}
\title{Information from the OpenMx model object}
\usage{
get_model_info(fit, covariates)
}
\arguments{
\item{fit}{fitted single-group RAM-type model.}

\item{covariates}{IPC regression predictor variables.}
}
\value{
List with model information.
}
\description{
Internal function that retrieves information from an OpenMx model object.
}
