% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moment_deviations.R
\name{get_moment_deviations}
\alias{get_moment_deviations}
\title{Calculate Individual Deviations from the Sample Moments}
\usage{
get_moment_deviations(x)
}
\arguments{
\item{x}{list with information about the model.}
}
\value{
Matrix with individual deviations from the sample mean and sample covariance.
}
\description{
Internal function that calculates the individual deviations from the sample mean and sample covariance.
}
