% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_updated_IPCs.R
\name{get_updated_IPCs}
\alias{get_updated_IPCs}
\title{Update the Individual Contributions to the Parameter Estimates}
\usage{
get_updated_IPCs(x)
}
\arguments{
\item{x}{list with information about the model.}
}
\value{
Matrix of initial individual parameter contributions.
}
\description{
Internal function that updates the individual contributions to the parameter estimates.
}
