% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weight_matrix.R
\name{get_weight_matrix}
\alias{get_weight_matrix}
\title{Calculate the Weight Matrix}
\usage{
get_weight_matrix(x)
}
\arguments{
\item{x}{list with information about the model.}
}
\value{
Estimated weight matrix.
}
\description{
Internal function that calculates the estimated weight matrix.
}
