% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weight_matrix2.R
\name{get_weight_matrix2}
\alias{get_weight_matrix2}
\title{Calculate the Weight Matrix}
\usage{
get_weight_matrix2(x, exp_cov_inv_up)
}
\arguments{
\item{x}{list with information about the model.}
}
\value{
Estimated weight matrix.
}
\description{
Internal function that calculates the estimated weight matrix.
}
