% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_definition_variables.R
\name{identify_definition_variables}
\alias{identify_definition_variables}
\title{Identifies definition variables in OpenMx models}
\usage{
identify_definition_variables(x)
}
\arguments{
\item{x}{An OpenMx model.}
}
\description{
This is an internal function that identifies which variables in the
data set stored in an \pkg{OpenMx} model are definition variables.
}
