% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcr2.R
\name{ipcr2}
\alias{ipcr2}
\title{Individual Parameter Contribution Regression (With Analytic Jacobian)}
\usage{
ipcr2(fit, covariates = NULL, iterated = FALSE, conv = 1e-04,
  max_it = 50)
}
\arguments{
\item{fit}{a structural equation model estimated with OpenMx. Only single-group RAM-type OpenMx model are allowed. The model must contain an estimated covariance structure. The mean structure is optional.}

\item{covariates}{a data frame or matrix with covariates.}

\item{iterated}{a logical value; if TRUE iterated IPC regression is performed after initial IPC regression.}

\item{max_it}{the maximum number of iterations for iterated IPC regressions.}

\item{convergence_criterion}{an integer used as stopping criterion for iterated IPC regression. Criterion is the largest difference in all parameter estimate between iterations.}
}
\value{
Returns a new ipcr object.
}
\description{
Explain parameter heterogeneity in a structural equation model with individual parameter contribution (IPC) regression. IPC Regression allows to test if parameter values differ across values of covariates.
}
\examples{
# Specify an OpenMx model
m <- mxModel(model = "CFA",
            manifestVars = c("x1", "x2", "x3"),
            latentVars = "f",
            type = "RAM",
            mxData(observed = ipcr_data, type = "raw"),
            mxPath(from = "f", to = c("x1", "x2", "x3"), arrows = 1,
                   free = c(FALSE, TRUE, TRUE), values = 1,
                   labels = c("l1", "l2", "l3")),
            mxPath(from = "f", arrows = 2, free = TRUE, values = 0.75,
                   labels = "var_f"),
            mxPath(from = c("x1", "x2", "x3"), arrows = 2, free = TRUE,
                   values = 0.25, labels = c("e1", "e2", "e3")),
            mxPath(from = "one", to = c("x1", "x2", "x3"), arrows = 1,
                   free = FALSE, values = 0))

# Fit the model
fit <- mxTryHard(model = m)

# Investigate the parameter estiamtes with IPC regression
IPC_reg <- ipcr(fit = fit, covariates = ipcr_covariates, iterated = TRUE)

# Get an overview about parameter differences
plot(IPC_reg)

# Show IPC regression output
summary(object = IPC_reg)
}
