% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcr_it.R
\name{ipcr_it}
\alias{ipcr_it}
\title{Iterated Individual Parameter Contribution Regression}
\usage{
ipcr_it(
  fit,
  covariates = NULL,
  iterate = FALSE,
  iteration_info = FALSE,
  conv = 1e-04,
  max_it = 50,
  regularization = FALSE,
  s = "lambda.min",
  alpha = 1,
  weights = NULL,
  nlambda = 100,
  standardize = TRUE,
  nfolds = 10,
  linear_MxModel = TRUE
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A number.
}
\description{
Performs iterated individual parameter contribution regression.
}
\examples{
add(1, 1)
add(10, 1)
}
