% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterated_ipcr.MxRAMModel.R
\name{iterated_ipcr.MxRAMModel}
\alias{iterated_ipcr.MxRAMModel}
\title{Iterated Individual Parameter Contribution Regression}
\usage{
\method{iterated_ipcr}{MxRAMModel}(x, IPC, iteration_info, covariates, conv, max_it, linear, ...)
}
\arguments{
\item{x}{a fitted model object}

\item{...}{arguments passed to methods.}
}
\value{
Returns a data frame.
}
\description{
This function performs iterated individual parameter contribution regression for fitted MxRAM-type models.
}
