% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterated_ipcr.R
\name{iterated_ipcr}
\alias{iterated_ipcr}
\title{Iterated Individual Parameter Contribution Regression}
\usage{
iterated_ipcr(x, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{...}{arguments passed to methods.}
}
\value{
Returns a list.
}
\description{
Generic function for computing iterated individual parameter contribution regression.
}
