% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.ipcr.R
\name{nobs.ipcr}
\alias{nobs.ipcr}
\title{Extract the Number of Observations from an ipcr Object}
\usage{
\method{nobs}{ipcr}(object, ...)
}
\arguments{
\item{object}{an ipcr object.}
}
\value{
A single integer.
}
\description{
Extract the number of observations from an ipcr object.
}
