% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobs.MxRAMModel.R
\name{nobs.MxRAMModel}
\alias{nobs.MxRAMModel}
\title{Extract the Number of Observations from a MxRAMModel Fit}
\usage{
\method{nobs}{MxRAMModel}(object, ...)
}
\arguments{
\item{object}{a fitted MXRAMModel-type model.}
}
\value{
A single integer.
}
\description{
Extract the number of observations from a MxRAMModel fit.
}
