% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ipcr.R
\name{predict.ipcr}
\alias{predict.ipcr}
\title{Predict Method for Linear Model Fits}
\usage{
\method{predict}{ipcr}(
  object,
  parameter = NULL,
  method = "iterated",
  newdata,
  se.fit = FALSE,
  scale = NULL,
  df = Inf,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  type = c("response", "terms"),
  terms = NULL,
  na.action = na.pass,
  pred.var = res.var/weights,
  weights = 1,
  ...
)
}
\arguments{
\item{object}{an ipcr object.}

\item{parameter}{a character vector with target model parameters that are to be inspected.}

\item{method}{a character string indicating whether to show results iterated IPC regression (iterated) or standard IPC regression (standard).}
}
\value{
A single integer.
}
\description{
Predicted values based on the IPC regression models.
}
