% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.ipcr.R
\name{residuals.ipcr}
\alias{residuals.ipcr}
\title{Extract Model Resiudals}
\usage{
\method{residuals}{ipcr}(object, parameter = NULL, method = "iterated", ...)
}
\arguments{
\item{object}{an ipcr object.}

\item{parameter}{a character vector with target model parameters that are to be inspected.}

\item{method}{a character string indicating whether to show results iterated IPC regression (iterated) or standard IPC regression (standard).}
}
\value{
Returns a list.
}
\description{
Extracts model residuals from IPC regression models.
}
