% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.ipcr.R
\name{sigma.ipcr}
\alias{sigma.ipcr}
\title{Extract Residual Standard Deviation 'Sigma'}
\usage{
\method{sigma}{ipcr}(object, parameter = NULL, method = "iterated", ...)
}
\arguments{
\item{object}{an ipcr object.}

\item{parameter}{a character vector with target model parameters that are to be inspected.}

\item{method}{a character string indicating whether to show results iterated IPC regression (iterated) or standard IPC regression (standard).}
}
\value{
A single integer.
}
\description{
Extract the estimated standard deviation of the errors from the IPC regression models.
}
