% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcr_reg.R
\name{summary.ipcr_reg}
\alias{summary.ipcr_reg}
\title{Summarize a Regularized IPCR Result}
\usage{
\method{summary}{ipcr_reg}(object, what = "everything", digits = 3, ...)
}
\arguments{
\item{object}{An object returned by \code{\link[=ipcr_reg]{ipcr_reg()}}.}

\item{what}{A single character string: \code{"everything"} (default),
\code{"predictors"}, or \code{"coefficients"}.}

\item{digits}{A nonnegative whole number of digits used for printing.}

\item{...}{Additional arguments passed to or from methods.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Summarize a Regularized IPCR Result
}
