% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary2.ipcr.R
\name{summary2.ipcr}
\alias{summary2.ipcr}
\title{Inidividual Parameter Contribution Regression Summary}
\usage{
summary2.ipcr(object, regularization = TRUE, digits = 3, verbose = FALSE, ...)
}
\arguments{
\item{object}{an ipcr object.}

\item{regularization}{a logical value; whether to show results for regularized IPC
regression (if computed) or non-regularized results.}

\item{digits}{integer indicating the number of decimal places to be used.}

\item{verbose}{a logical value; if TRUE ipcr settings, non-regularized, and regularized
results are shown.}

\item{...}{further arguments passed to and from methods.}
}
\description{
This functions returns the coefficients of the individual parameter
contribution (IPC) regression equations.
}
