% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.ipcr.R
\name{vcov.ipcr}
\alias{vcov.ipcr}
\title{Calculate Variance-Covariance Matrices for the IPC Regression Models}
\usage{
\method{vcov}{ipcr}(object, parameter = NULL, method = "iterated", ...)
}
\arguments{
\item{object}{an ipcr object.}

\item{parameter}{a character vector with target model parameters that are to be inspected.}

\item{method}{a character string indicating whether to show results iterated IPC regression (iterated) or standard IPC regression (standard).}
}
\value{
A single integer.
}
\description{
Returns the variance-covariance matrix of the main parameters of the IPC regression models. The “main” parameters of the model correspond to those returned by coef, and do not contain a nuisance scale parameter (sigma).
}
