% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipcs.R
\name{get_ipcs}
\alias{get_ipcs}
\title{Calcuate Individual Parameter Contributions}
\usage{
get_ipcs(x)
}
\arguments{
\item{x}{a fitted model object.}
}
\value{
Returns a data.frame containing the individual parameter contributions of a
fitted model.
}
\description{
This functions calculates the individual parameter contributions of a
fitted model.
}
