% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores}
\alias{get_scores}
\title{Extracts Scores}
\usage{
get_scores(x, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{...}{arguments passed to methods.}
}
\value{
Returns a matrix containing the scores of a fitted model.
}
\description{
This functions extracts the scores of a fitted model. Arguments are passed
to \link[sandwich]{estfun}.
}
