% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcr_reg.R
\name{ipcr_reg}
\alias{ipcr_reg}
\title{Add together two numbers}
\usage{
ipcr_reg(
  fit,
  covariates = NULL,
  iterate = FALSE,
  iteration_info = FALSE,
  conv = 1e-04,
  max_it = 50,
  regularization = FALSE,
  s = "lambda.min",
  alpha = 1,
  weights = NULL,
  nlambda = 100,
  standardize = TRUE,
  nfolds = 10,
  linear_MxModel = TRUE
)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A number.
}
\description{
Add together two numbers
}
