% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ipcr}
\alias{plot.ipcr}
\title{Plot Correlations Between Predictors and IPCs}
\usage{
\method{plot}{ipcr}(x, print_corr = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"ipcr"}.}

\item{print_corr}{Logical; if \code{TRUE}, correlation coefficients are
displayed on the heatmap. Default is \code{FALSE}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{ggplot} object.
}
\description{
Generates a heatmap visualizing the correlations between
predictors and individual parameter contributions (IPCs).
}
\details{
This function is a wrapper for \code{\link[ggplot2]{ggplot}}.
Factor and logical predictors are represented by the columns of the design
matrix used for the IPC regression. Correlations involving a constant IPC or
predictor column are undefined and are displayed in grey.
}
