% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ipcr.R
\name{plot.ipcr}
\alias{plot.ipcr}
\title{Plot Correlations between Covariates and IPCs}
\usage{
\method{plot}{ipcr}(x, ...)
}
\arguments{
\item{x}{an ipcr object.}

\item{...}{other arguments.}
}
\description{
This functions plots a heat map that visualizes the correlation between
covariates and IPCs.
}
\details{
This function is a wrapper for \code{\link[ggplot2]{ggplot}}. Currently,
arguments passed to \code{ggplot} cannot be changed.
}
