% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ipcr.R
\name{plot.ipcr}
\alias{plot.ipcr}
\title{Plot Correlation between Covariates and IPCs}
\usage{
\method{plot}{ipcr}(x, ipc_method = "initial", cor_method = "pearson")
}
\arguments{
\item{x}{an ipcr object.}

\item{ipc_method}{character string indicating whether the relationship between covariates and initial IPCs (method = “initial”, default option) or the iterated IPCs (method = “iterated”) is drawn.}

\item{cor_method}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman".}
}
\description{
This functions plots a heat map that visualizes the correlation between covariates and IPCs.
}
\examples{
# Plot the correlation between covariates and initial IPCs.
plot(x = IPC_reg)
}
