% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differences.R
\name{plot_differences}
\alias{plot_differences}
\title{Plot Estimated Conditional Differences in Model Parameters}
\usage{
plot_differences(
  x,
  parameter = NULL,
  predictor = NULL,
  confidence_level = 0.95
)
}
\arguments{
\item{x}{An object of class \code{"ipcr"}.}

\item{parameter}{A character string specifying the name of a model parameter
as displayed by \code{print}. By default, all model parameters are plotted.}

\item{predictor}{A character string specifying the name of a predictor.
By default, the effects of all predictors are plotted.}

\item{confidence_level}{A numeric value indicating the confidence level
for the confidence intervals. The default is \code{0.95}, corresponding
to a 95 percent confidence interval.}
}
\description{
Generates plots of estimated model parameters as a function of
one or more predictors.
}
\details{
The function plots regression lines for continuous predictors and group means for
dummy variables. Confidence intervals are included based on the model parameter
estimates and their corresponding variances and covariances. The plots visualize
the estimated parameter values as a function of a predictor, assuming all other
predictors are set to zero. Mean-centering predictors may improve interpretability.
This function is a wrapper for \code{\link[ggplot2]{ggplot}}.
}
\examples{
# Generate data
## predictors
z1 <- rep(0:1, each = 50)
z2 <- rnorm(n = 100)
predictors <- data.frame(z1 = z1, z2 = z2)
## Model data
x <- rnorm(n = 100)
y <- 0.5 + 0.75*z1 + rnorm(n = 100, sd = sqrt(0.75))
d <- data.frame(x = x, y = y)

# Fit a linear regression
m <- lm(y ~ x, data = d)

# Investigate model with IPC regression
res <- ipcr(m, predictors = predictors)

# Plotting parameter values as a function of the predictors
## All plots
plot_differences(res)
## Plot the values of the regression slope x as a function of the predictor z1
plot_differences(res, parameter = "x", predictor = "z1")

}
\seealso{
\code{\link{plot.ipcr}}
}
