% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ipcr.R
\name{print.ipcr}
\alias{print.ipcr}
\title{Print ipcr Object Information}
\usage{
\method{print}{ipcr}(x)
}
\arguments{
\item{object}{an ipcr object.}
}
\description{
This function prints the some of the main information of an ipcr object.
}
\examples{
# Print main information
print(x = IPC_reg)
}
