% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.summary.R
\name{summary.ipcr}
\alias{summary.ipcr}
\title{Inidividual Parameter Contribution Regression Summary}
\usage{
\method{summary}{ipcr}(object, parameter = NULL, method = "both")
}
\arguments{
\item{object}{an ipcr object.}

\item{parameter}{a character vector with the target parameters of the structural equation model (SEM) that are to be inspected.}

\item{method}{a character string indicating the method to be used; for method = "both" static and iterated IPC regression is used (default), for method = "static" only static IPC coefficients are displayed, and for method = "iterate" only iterated coefficients are shown.}
}
\description{
This functions returns the estimated coefficients of the individual parameter contribution (IPC) regression equations as well as the corresponding standard errors, t-values, and p-values.
}
\examples{
# Display the iterated individual contribution regression parameters of all SEM parameters.
summary(object = IPC_reg, method = "iterate")
}
